+++ src/main/java/fish/global/config/WebSocketHandler.java
@@ -1,5 +1,7 @@
 package fish.global.config;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.CloseStatus;
 import org.springframework.web.socket.TextMessage;
@@ -12,14 +14,17 @@ import java.util.concurrent.ConcurrentHashMap;
 @Component
 public class WebSocketHandler extends TextWebSocketHandler {
     private static final ConcurrentHashMap<String, WebSocketSession> CLIENTS = new ConcurrentHashMap<>();
+    private static final Logger log = LoggerFactory.getLogger(WebSocketHandler.class);
 
     @Override
     public void afterConnectionEstablished(WebSocketSession session) {
+        log.info("Session Connection Established!✅ :: {}", session.getId());
         CLIENTS.put(session.getId(), session);
     }
 
     @Override
     public void afterConnectionClosed(WebSocketSession session, CloseStatus status) {
+        log.info("Session Connection Closed!👊 :: session id = {}, close status = {}", session.getId(), status);
         CLIENTS.remove(session.getId());
     }
